+++ src/cpe/utils/tests-ut/test_trunk.cpp
@@ -5,8 +5,8 @@ TEST_F(BufferTest, trunk_create) {
     struct mem_buffer_trunk * trunk = 
         mem_buffer_append_trunk(&m_buffer, 3);
 
-    EXPECT_EQ(3, trunk->m_capacity);
-    EXPECT_EQ(0, trunk->m_size);
+    EXPECT_EQ((size_t)3, trunk->m_capacity);
+    EXPECT_EQ((size_t)0, trunk->m_size);
     EXPECT_TRUE(trunk->m_allocrator == m_buffer.m_default_allocrator);
 }
 
@@ -16,9 +16,9 @@ TEST_F(BufferTest, trunk_append_basic) {
 
     const char * data = "abcde";
 
-    EXPECT_EQ(3, mem_trunk_append(&m_buffer, trunk, data, 3));
-    EXPECT_EQ(3, trunk->m_size);
-    EXPECT_EQ(3, mem_buffer_size(&m_buffer));
+    EXPECT_EQ((size_t)3, mem_trunk_append(&m_buffer, trunk, data, 3));
+    EXPECT_EQ((size_t)3, trunk->m_size);
+    EXPECT_EQ((size_t)3, mem_buffer_size(&m_buffer));
 }
 
 TEST_F(BufferTest, trunk_append_twice) {
@@ -27,15 +27,15 @@ TEST_F(BufferTest, trunk_append_twice) {
 
     const char * data = "abcde";
 
-    EXPECT_EQ(3, mem_trunk_append(&m_buffer, trunk, data, 3));
-    EXPECT_EQ(3, mem_trunk_append(&m_buffer, trunk, data, 3));
-    EXPECT_EQ(6, trunk->m_size);
-    EXPECT_EQ(6, mem_buffer_size(&m_buffer));
+    EXPECT_EQ((size_t)3, mem_trunk_append(&m_buffer, trunk, data, 3));
+    EXPECT_EQ((size_t)3, mem_trunk_append(&m_buffer, trunk, data, 3));
+    EXPECT_EQ((size_t)6, trunk->m_size);
+    EXPECT_EQ((size_t)6, mem_buffer_size(&m_buffer));
 
     char buf = 0;
-    EXPECT_EQ(1, mem_trunk_append(&m_buffer, trunk, &buf, 1));
-    EXPECT_EQ(7, trunk->m_size);
-    EXPECT_EQ(7, mem_buffer_size(&m_buffer));
+    EXPECT_EQ((size_t)1, mem_trunk_append(&m_buffer, trunk, &buf, 1));
+    EXPECT_EQ((size_t)7, trunk->m_size);
+    EXPECT_EQ((size_t)7, mem_buffer_size(&m_buffer));
 
     EXPECT_STREQ("abcabc", as_string());
 }
@@ -46,9 +46,9 @@ TEST_F(BufferTest, trunk_append_overflow) {
 
     const char * data = "abcde";
 
-    EXPECT_EQ(2, mem_trunk_append(&m_buffer, trunk, data, 3));
-    EXPECT_EQ(2, trunk->m_size);
-    EXPECT_EQ(2, mem_buffer_size(&m_buffer));
+    EXPECT_EQ((size_t)2, mem_trunk_append(&m_buffer, trunk, data, 3));
+    EXPECT_EQ((size_t)2, trunk->m_size);
+    EXPECT_EQ((size_t)2, mem_buffer_size(&m_buffer));
 }
 
 TEST_F(BufferTest, trunk_append_twice_overflow) {
@@ -57,14 +57,14 @@ TEST_F(BufferTest, trunk_append_twice_overflow) {
 
     const char * data = "abcde";
 
-    EXPECT_EQ(3, mem_trunk_append(&m_buffer, trunk, data, 3));
-    EXPECT_EQ(3, trunk->m_size);
-    EXPECT_EQ(3, mem_buffer_size(&m_buffer));
+    EXPECT_EQ((size_t)3, mem_trunk_append(&m_buffer, trunk, data, 3));
+    EXPECT_EQ((size_t)3, trunk->m_size);
+    EXPECT_EQ((size_t)3, mem_buffer_size(&m_buffer));
 
 
-    EXPECT_EQ(2, mem_trunk_append(&m_buffer, trunk, data, 3));
-    EXPECT_EQ(5, trunk->m_size);
-    EXPECT_EQ(5, mem_buffer_size(&m_buffer));
+    EXPECT_EQ((size_t)2, mem_trunk_append(&m_buffer, trunk, data, 3));
+    EXPECT_EQ((size_t)5, trunk->m_size);
+    EXPECT_EQ((size_t)5, mem_buffer_size(&m_buffer));
 }
 
 TEST_F(BufferTest, trunk_free) {
@@ -73,12 +73,12 @@ TEST_F(BufferTest, trunk_free) {
 
     const char * data = "abcde";
 
-    EXPECT_EQ(3, mem_trunk_append(&m_buffer, trunk, data, 3));
-    EXPECT_EQ(3, trunk->m_size);
-    EXPECT_EQ(3, mem_buffer_size(&m_buffer));
+    EXPECT_EQ((size_t)3, mem_trunk_append(&m_buffer, trunk, data, 3));
+    EXPECT_EQ((size_t)3, trunk->m_size);
+    EXPECT_EQ((size_t)3, mem_buffer_size(&m_buffer));
 
     mem_trunk_free(&m_buffer, trunk);
-    EXPECT_EQ(0, mem_buffer_size(&m_buffer));
+    EXPECT_EQ((size_t)0, mem_buffer_size(&m_buffer));
 }
 
 TEST_F(BufferTest, trunk_free_middle) {
@@ -92,8 +92,8 @@ TEST_F(BufferTest, trunk_free_middle) {
         mem_buffer_append_trunk(&m_buffer, 5);
     mem_trunk_append(&m_buffer, trunk2, data, 4);
 
-    EXPECT_EQ(7, mem_buffer_size(&m_buffer));
+    EXPECT_EQ((size_t)7, mem_buffer_size(&m_buffer));
 
     mem_trunk_free(&m_buffer, trunk2);
-    EXPECT_EQ(3, mem_buffer_size(&m_buffer));
+    EXPECT_EQ((size_t)3, mem_buffer_size(&m_buffer));
 }
